[FIX] delivery_envia: fix error on fetching shipping rate without geolocation
Currently, an error is raised when trying to fetch the shipping rate if the
partner does not have the required geolocation fields (country, state, and city).

Steps to Reproduce:
1. Install and configure the Envia Shipping module.
2. Create a partner without an address (only name + phone).
3. Create quotation for that partner with a deliverable product (e.g; Conference Chair).
4. Click "Add Shipping", choose Envia Shipping Method, and then click "Get Rate".

Error:
TypeError - quote_from_bytes() expected bytes

Cause:
At [1], the system tries to compute Envia shipping rates based on the partner’s
country, state, and city. If any of these fields are not set, an error is raised.

Fix:
This commit adds a check for the required fields (country, state, and city).
If any are missing, `_geolocate_zip` returns False, leading to a proper
validation error instead of a traceback. - [2]

[1]: https://github.com/odoo/enterprise/blob/f1a02626a1fbe76add104832e151c647307f3ae7/delivery_envia/models/envia_request.py#L591-L593
[2]: https://github.com/odoo/enterprise/blob/f1a02626a1fbe76add104832e151c647307f3ae7/delivery_envia/models/envia_request.py#L617-L624

sentry-7063870478

closes odoo/enterprise#101765

X-original-commit: 44095df6ba3d95aace5f2f00359073fc5a1dd59e
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>
Signed-off-by: Aditi Patel (adip) <adip@odoo.com>