[FIX] hr_holidays_gantt: prevent error when working schedule is not set
Currently, an error occurs when an employee has no working schedule set and the
user tries to open the Time Off overview in the Gantt view.

Steps to Reproduce:
1. Install `hr_holidays_gantt` with demo data.
2. Remove the Working Hours(Payroll section) of Marc Demo.
3. Navigate to: Time off > Overview.

Error:
AttributeError -  'bool' object has no attribute 'upper'

Cause:
The error occurs because at [1], the contract does not have a resource_calendar_id
(working schedule) assigned. As a result, it attempts to access attributes of
a False value, causing the error.

Fix:
This commit skips the computation when the contract does not have a working
schedule set.

[1]: https://github.com/odoo/enterprise/blob/db049f42ad4c2c291b2d64a9e637cd2292c280e8/hr_holidays_gantt/models/hr_leave.py#L204

sentry-7017326141

closes odoo/enterprise#100955

X-original-commit: 7506cfcea97419f9eadbc0b916805bf7eeccec24
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>