[FIX] mrp_workorder: await records update in test_shop_floor_my_wo_filter_with_pin_user
The test `test_shop_floor_my_wo_filter_with_pin_user` sometimes fails
on these steps:
https://github.com/odoo/enterprise/blob/422ac3d5b10c44233321010b9ecb8e37735b3ea3/mrp_workorder/static/tests/tours/tour_shopfloor.js#L177-L190
https://github.com/odoo/enterprise/blob/422ac3d5b10c44233321010b9ecb8e37735b3ea3/mrp_workorder/static/tests/tours/tour_shopfloor.js#L196-L206
https://github.com/odoo/enterprise/blob/422ac3d5b10c44233321010b9ecb8e37735b3ea3/mrp_workorder/static/tests/tours/tour_shopfloor.js#L212-L221
This happends since changing the user requires some time to display the
related shopfloor records, but the steps check the number of visible
records as soon as it has switched rather than when it is sure that the
records are displayed.

Fix:

Since switching employees will first empty the recordset and later
display the related records, we can split the steps in two. We first
check that we switched users, then we check the existence of a record
that is not present for the previous user, and only then perform the
related checks.

runbot-226734

closes odoo/enterprise#101602

X-original-commit: 3a1d49174116137b3938a5ebaee5e80e184f15fd
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>
Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>