[FIX] account_accountant: move test case to enterprise
Community build was failing with:
```test_bill_state_change_on_payment_state
    self.assertEqual(payment.invoice_ids.payment_state, 'not_paid')
AssertionError: 'paid' != 'not_paid'
- paid
+ not_paid
```

In community there is no any `in_payment` state for account move records. As on
changing payment state to draft It stayed in the `paid` only. `in_payment` state
introduced in the enterprise module. So the test case is failing for the
community version.

To fix this I've moved the test case to the enterprise to retain the expected
behaviour.

runbot error: 234453

closes odoo/enterprise#100922

X-original-commit: 763b34029253a786d1e4bf09a780c1fb799f0ba5
Related: odoo/odoo#238034
Signed-off-by: Paolo Gatti (pgi) <pgi@odoo.com>
Signed-off-by: Joyal Patel (pajo) <pajo@odoo.com>