[FIX] documents: show "Restricted" when no access on the folder
Bug
===
Since 8cb1cf5c490e6c29edc54b199b3a61ca42bf17bd , we never show
"Restricted" in the kanban view of documents, when we don't have
access on the folder.

Reason
======
For AI, because we didn't want to add "Restricted Folder" in the prompt
when the user has no access on the folder, we skipped that in the compute
of the display_name when we are in sudo. But display_name are always
computer in sudo, so we never show "Restricted Folder" in the kanban
view of documents.

Expected
========
When we don't have access on the folder of a document,
we want to show "Restricted Folder" in the kanban view.

We want the LLM to get the name of the parent folder even if the
current user has no access, but in documents, we want
to show "Restricted",

Task-5362100

closes odoo/enterprise#100850

Signed-off-by: Patrick Hoste (pko) <pko@odoo.com>