[FIX] accountant_knowledge: call custom `export_to_pdf` if exists
Currently there's a traceback when trying to print an audit report.

Steps to reproduce:
- Go to "Accounting/Review/Audit/Annual Report" and create an Audit
Report.
- click "Print"
-> Internal Server Error

This happens when generating the Journal Audit pdf, this report needs
to uses the `export_to_pdf` that is defined in its custom handler
`account.journal.report.handler`.

The solution is to use `dispatch_report_action`, as it will determine
which function to use in order to export the pdf for each report.

task-5154292

closes odoo/enterprise#100949

Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>