[IMP] account_accountant: open list view for invalid statements alert
When the dashboard shows an invalid statement, clicking the alert redirected
users to the kanban view of statements. That view does not allow editing or
performing any corrective actions, forcing users to manually switch to the list
view to fix the invalid entries.

This commit updates the 'Invalid Statement(s)' alert so it redirects directly to
the list view, allowing users to correct the statements without extra navigation.

task-5346063

closes odoo/enterprise#100407

Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>