[FIX] l10n_cz_reports: fix VAT control statement
With l10n_cz_reports:

- Create a currency exchange between CZK and EUR where the EUR is valued at least at twice the amount of CZK.
- Create an invoice in EUR, with a line with price_unit 5000 and a tax.
- In the CZ Tax Report, in the VAT control statement, the converted amount is found in section B.3, which contains received taxable supplies and provided payments up to CZK 10,000. However, the converted amount of the invoice in CZK is higher than 10,000.

In `_report_custom_engine_control_statement`, the amount used to check whether the move should be included in this section uses `amount_total`, which in the case of foreign currency gives the wrong result. If the move is in a foreign currency the total is not in CZK so we have to use the absolute value of the signed total.

opw-5080339

closes odoo/enterprise#101730

X-original-commit: 9cbb81fdfef6c23fc553a4530cf6b0efcdecad0c
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Guillaume Teboul-Tornezy (gute) <gute@odoo.com>