[FIX] pos_l10n_se: allow price control with blackbox
Before this commit, user couldn't control the price in the POS if using the swedish blackbox.

After this commit, user can control the price.

It's not clear why the behavior at integration was set to this but it appears that it's not mandatory for swedish blackbox unlike the belgian one.

opw-5253401

closes odoo/enterprise#102177

X-original-commit: 9cdc314eadfde9bbe3fa6aee361b77f4cc0f6278
Signed-off-by: Max Whale (mawh) <mawh@odoo.com>