[FIX] pos_settle_due: settle correctly credit notes
We had a bug when settling invoices and credit notes of a customer.
The credit notes where not correctly computed.

Steps to reproduce:
-------------------
In accounting:
* Create and confirm a customer invoice for a total of 10$.
* Create and confirm for the same customer a credit note for a total of 2$.
In POS:
* In a seesion, open the customer selection menu.
* In the burger menu at the right of our customer, select Settle invioces.
* Select our invoice and credit note.

> Observation:
The credit note was previously seen as an amount to pay.

Why the fix:
------------
Recomputing updates existing credit notes to use the signed residual logic.
The domain change allows credit notes with negative pos_amount_unsettled to appear.
After these changes, credit notes should appear in the "Settle invoices" dialog with negative amounts, and selecting them will create negative lines that reduce the total.

opw-5257884

closes odoo/enterprise#102052

X-original-commit: 76b47ce2dd2d7f2e722b1bea3069a15efc27da1f
Signed-off-by: Stéphane Vanmeerhaeghe (stva) <stva@odoo.com>
Signed-off-by: Guillaume Vardakis (gvar) <gvar@odoo.com>