[FIX] hr_payroll: write full name on payslip generated from payrun
_
Short functional explanation of the error:
When generating a payslip from a payrun for an employee, the name
of the PDF payslip is simply "new payslip" instead of
"Salary Slip - <name of employee> - <time period of the slip>"

Reproduction Steps:
1. Go to Payroll and click on the Payslips tab > payslips.
2. Click on Pay Run. Select Regular Pay and click Continue.
3. Select an employee for which you'd like to generate the payslip.
4. Click on the employee row in the list view.
5. Click Compute sheet > Print.

 Expected behavior:
A PDF with name "Salary Slip - <employee name> - <time period>" is
generated.

Unexpected Behavior:
A PDF with name "new Payslip" is generated.

Origin of the issue:
When printing the slip with this flow, we don't call the method
```_compute_name()``` used to compute the name of the current slip.
Therefore, it stays at 'New Slip', which is the default name.
Henceforth, we have to call this method manually when printing the
slip.
_
opw-5216796

---
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr

closes odoo/enterprise#101259

X-original-commit: e317322b3699a8f3d570d4613c17625fb9a0d785
Signed-off-by: Mathias Lebel (lebm) <lebm@odoo.com>
Signed-off-by: Maeva de Keyser (mdek) <mdek@odoo.com>