[FIX] l10n_uy_edi: assign the computed fields for new doc record
The PR https://github.com/odoo/odoo/pull/209587 adds a check for
onchange calls on a newly created record. Since in that case
`move_id` is not set on `l10n_uy.document`, the compute method
fails to assign values.

This commit adds a pre-assigned fallback for this specific case.

error-234425

closes odoo/enterprise#101113

X-original-commit: a7f1dfe1a35652f7ecf2c7381475fac093559ef7
Signed-off-by: Laurent Smet (las) <las@odoo.com>