[FIX] web_studio: correctly autofocus new button in form view
Have an arch with many conditional button in the header
Click on Add a button in the form editor.

Before this commit, the new button was not correctly focused and the sidebar did not display the properties of the new button.
This was because we gave the wrong xpath coordinates to focus to the editor.

After this commit, this flow works as expected

task-5364798

closes odoo/enterprise#102380

X-original-commit: 3ce1a868e94b4e63fefa0035bdac9abd561e2520
Signed-off-by: Hubert Van De Walle <huvw@odoo.com>
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>