[FIX] spreadsheet_edition: avoid crash on filter creation
This is a complement to #72576.
The mentioned PR allowed the users to go through the datasources
sidepanels to fix them after a migration that would make there source
model invalid (renaming/deletion).

However, we did not account for the edition of all filters. If a user
wanted to create a text or a date filter on a spreadsheet with broken
datasources, they would be greeted with a traceback.

This revision extends the failsafe of the relational filters to all the
filters.

task-5220489

Part-of: odoo/enterprise#100989
Related: odoo/odoo#238155
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>