[FIX] account_accountant: Show Reviewed button
Before this commit, we called the wrong variable
to know whether the "Reviewed" button is available
or not. It leads to have a no show of this button
even when it was needed.

This commit also adds a small security check to
remove the secondary buttons when the statement is
reconciled.

This commit also changes all the usage of
props.statementLine.data for statementLineData getter.

no task id

closes odoo/enterprise#101063

Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>