[FIX] spreadsheet_edition: prevent relational filter panel crash
When a relational global filter is invalid (e.g. its target model does
not exist), we can still edit it and try to edit its default value.
clicking on the related input will traceback.
Same goes with the boolean field to add a subdomain to the filter
values.

This commit hides those two options when the filter is invalid.

task-5220489

closes odoo/enterprise#100989

Related: odoo/odoo#238155
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>