[FIX] sale_commission: achievement view was not available in session
In the commission module, the achievements are computed with a temporary SQL
view. These view have limited visibility inside a SQL session. Before
this commit, we checked the existence of a temporary view with the
following SQL request:

SELECT viewname FROM pg_catalog.pg_views WHERE viewname='sale_commission_achievement_report_view'

This simple query would return the existence of the view but not its
availability in the current session. As a result, when the user tried to
access the report values,

SELECT * FROM sale_commission_achievement_report_view

would result in an error as the temporary view would not be available
in all the running sessions. Only the one which created the temporary
view would be able to access the data.

After this commit, the existence and the availability of the view is
checked to make sure that it is possible to fetch the results in
the current session.

task-5378382

closes odoo/enterprise#101070

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>