[FIX] ai: transcription summary in proper language
This commit introduces a change that makes the agent produce a summary
in the selected language from the transcription tab.

It injects a specific prompt using the `currentLanguage` of the
transcription the component. Instructing the LLM to generate in the
proper language.

The dates inserted are now in the language selected of the transcription
component as well.

closes odoo/enterprise#101075

Signed-off-by: Panagiotis Kyriakou (paky) <paky@odoo.com>