[FIX] account_online_synchronization: prevent error when no journal is linked
Currently, an error occurs when user opens the
account.duplicate.transaction.wizard.form view and saves it.

Steps to Reproduce:

 - Install the account_accountant module.
 - Using Open View, open the account.duplicate.transaction.wizard view > enter
   the start date, and save..

ValueError: Compute method failed to assign account.duplicate.transaction.wizard(22,).provider_duplicate_ids

The error occurs because when the view is opened and saved, the
_compute_provider_duplicate_ids method is triggered. Since no journal
is linked to the wizard, the method assigns an empty list ([]) to
provider_duplicate_ids [1]. Because this is a O2M field, assigning a raw
empty list is invalid and causes the error.

This commit ensures that when journal is missing, instead of assigning
an empty list ([]), a valid relational command ([Command.set([])]) is used.

Ref: https://github.com/odoo/enterprise/pull/97196/commits/c109e0bcce20f14b6dbd3d28e61b135f1fd048f2

[1]: https://github.com/odoo/enterprise/blob/d808ad3f6ff52632e00cf1a68de0e5270e52895d/account_online_synchronization/wizard/account_journal_duplicate_transactions.py#L53

sentry-7080686084

closes odoo/enterprise#102187

X-original-commit: db6031e383d39dd690f5d2958ff18e06f53a7c41
Signed-off-by: Paolo Gatti (pgi) <pgi@odoo.com>