[FIX] l10n_be_hr_payroll_partena: input export's company code in CSV file
Issue:
In multicompany, when we generate the Partena export file of the 'not active' company,
the partena code of the active company is inputted in the file.

Steps to reproduce:
- Install 'l10n_be_hr_payroll_partena' and switch to a Belgian company
- Make sure the company has a "Partena Affiliation Number"
- Create an employee for this company, with a "Partena code"
- Create a contract for this employee, set it a running
- Create a new Belgian company with a different "Partena Affiliation Number"
- Activate both Belgian companies, but set the second one as active
- Payroll > Reporting > Export work entries to Partena
- Create a new one, populate it with the employee just created
- Click "Generate Export File"

Cause:
When getting the data for the CSV file, we use `self.env.company` which is the active company.
So when this company is not the one of the export record, we input the wrong code values.

Solution:
Use `self.company_id` instead of `self.env.company_id`.

Also adds the test class with basic tests.

opw-5345786

closes odoo/enterprise#102062

X-original-commit: 9dfdf185e019cbf7bf6416d0eae2c70c5237a60b
Signed-off-by: Mélanie Peyrat (mepe) <mepe@odoo.com>
Signed-off-by: Mathieu Coutant (mcou) <mcou@odoo.com>