[FIX] account_reports: Use the main company when getting the report options
When calling get_options, we are giving the companies from the tax unit.
The main company is the first element in `allowed_company_ids`.
However, we cannot be sure of the order of the list of companies from the tax unit, so we need to specify it with `with_company`.

X-original-commit: 9103618ace52285e20c949b8d8738a050cbcbd6d
Part-of: odoo/enterprise#101285
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Léo Leclerc (leol) <leol@odoo.com>