[FIX] l10n_cl_edi_pos: fix TypeError when refunding orders with Consumidor Final Anónimo
We had a bug when attempting to refund orders that were created with the
"Consumidor Final Anónimo" customer. The refund validation would crash
with a TypeError instead of showing the proper error message.

Steps to reproduce:
-------------------
In POS with l10n_cl_edi module activated:
* Create a new order with "Consumidor Final Anónimo" as customer
* Add products and pay the order
* Validate the order
* Attempt to refund this order

> Observation:
The refund validation would crash with:
TypeError: Cannot read properties of undefined (reading 'add')
at OrderPaymentValidation.isOrderValid

Why the fix:
------------
The code was trying to access `this.dialog` which is undefined in the
OrderPaymentValidation class context. The dialog service should be accessed
via `this.pos.dialog`, which is the correct pattern used throughout the
base OrderPaymentValidation class. This fix ensures the error dialog is
properly displayed when attempting to refund orders for the anonymous
final consumer, instead of crashing with a TypeError.

closes odoo/enterprise#101356

Signed-off-by: Stéphane Vanmeerhaeghe (stva) <stva@odoo.com>