[FIX] l10n_ke_edi_oscu: prevent error when sending vendor bill to eTIMS
Currently, an error occurs when clicking the "Send to eTIMS" button on vendor
bills.

Steps to Reproduce:
 - Install 'l10n_ke_edi_oscu' module without demo data.
 - Create a New company with Kenya as the country and switch to it.
 - Go to Vendors > Bills, create a new bill, and add an invoice line without tax.
 - Confirm it, then click 'Send to eTIMS'.

Traceback:
 File "/home/odoo/src/enterprise/19.0/l10n_ke_edi_oscu/models/account_move.py", line 551, in action_l10n_ke_oscu_confirm_vendor_bill
    content = move._l10n_ke_oscu_json_from_move()
  File "/home/odoo/src/enterprise/19.0/l10n_ke_edi_oscu/models/account_move.py", line 199, in _l10n_ke_oscu_json_from_move
    line_items = self._l10n_ke_oscu_get_json_from_lines(tax_details)
  File "/home/odoo/src/enterprise/19.0/l10n_ke_edi_oscu/models/account_move.py", line 250, in _l10n_ke_oscu_get_json_from_lines
    tax, line_tax_details = next(
StopIteration: null

This error occurs because when a new company is created, the 'eTIMS Server Mode'
in Settings is empty. As a result, 'l10n_ke_oscu_is_active' field becomes
'False', and at [1] the 'l10n_ke_validation_message' field will also be 'False',
causing the flow to be skipped. Therefore, no error is raised on the frontend
side. The field 'l10n_ke_oscu_is_active' is set to 'True' only when
'eTIMS Server Mode' is set to ‘Demo’.

Here we raise a warning when the eTIMS configuration is not set up correctly.

[1]: https://github.com/odoo/enterprise/blob/7fb7b3168039f00b6d815202bc19ac35aa1d9b5e/l10n_ke_edi_oscu/models/account_move.py#L91-L93

sentry-7083978544

closes odoo/enterprise#102600

X-original-commit: 8c78236b63f8fa7f9209d4365214a84e05e1cd42
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Aayush Modi (aamo) <aamo@odoo.com>