[FIX] l10n_be_hr_payroll: fix issue in time off allocation wizard
this commit fixs an issue introduced on production due to migration to
versioned employees by adding a where clause to select employees with
their versions belonging to the selected companies only. This prevents
from accessing employees from other companies later when summing the
attestation days that might not be accessed by the user.

task-5386951

closes odoo/enterprise#101517

X-original-commit: 406390e6a98088ee96bbd0ae74f91124bfafd4be
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: Ahmed Taha Abdelaziz (ahta) <ahta@odoo.com>