[FIX] quality_control: do not delete QC for unfinished pickings
When splitting a picking (thus creating a backorder), the quality
checks that are still in state 'none' are deleted because Odoo
assumes that the old picking is done and the QCs are not needed
anymore.

This fix ensures that QCs for split pickings that are still in
progress remain.

opw-5193424

closes odoo/enterprise#101532

X-original-commit: 0844194292437cc0bc243603f8622f836998c28e
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>
Signed-off-by: Stéphane Diez (snd) <snd@odoo.com>