[FIX] ai: chatgpt plugin toolbar button bug
This commit fixes an issue with the toolbar button. Before when no
selection were applied in the HTML Editor, the button would be disabled.

This commit introduce a fix that makes the button enabled in two
scenarios:
- Either the user has no selection, clicking the button will then open
the chat window without any AI button, as it does with the powerbox
buttons,
- Or the user has a selection, then clicking the button will result in
opening the chat window with the AI buttons as it was already the case.

If the user has a selection that is empty, then the button remains
visible but is disabled.

task-5220018

closes odoo/enterprise#101332

Signed-off-by: Panagiotis Kyriakou (paky) <paky@odoo.com>