[FIX] stock_barcode: open list view on mobile
Issue
=====
On mobile, we should prioritize kanban views over list views because
kanban views are usually more suitable for small device screen.
That said, when a product's barcode is scanned in the Barcode app main
menu, we show this product's stock locations but we do that with a list
view, no matter if the user is on a big screen or a small screen.

How to reproduce
================
On mobile device:
- Enable location and have a product with a barcode and with quantities
  in two different locations;
- Open Barcode app;
- Scan the product's barcode => The product's stock locations are
  displayed in a list view, which is not very pratical on small device.

Fix
===
The action key `mobile_view_mode` was not set, with this key, we can
define what view type we want to prioritize for mobile device.
This commit partially reverts 4aead7febcf76694a094fc676ebb0e5db61c9acb
(the part which displays always the kanban view no matter the device)
but keeps changes done in the tour and run two times the tour:
- One time as usually (large screen device), checking products are
  displayed as line;
- One time as a mobile device (small screen), checking products are
  displayed as kanban record.

opw-5180783

closes odoo/enterprise#101957

X-original-commit: 9b19b77bb545a07bb654fe3830f43ea9c4902eaa
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Steve Van Essche <svs@odoo.com>