[FIX] l10n_be_reports: remove wrong grid number from VAT return XML
Steps to reproduce:
- Install accountant and l10n_be_reports
- Switch to a Belgian company (e.g. BE Company CoA)

- Create an Intra-Community invoice:
  * Customer: [EU customer]
  * Invoice Date: [last month]
  * Fiscal Position: [Intra-Community]
  * Invoice Lines: [a product with "0% EU M" tax]
- Confirm the invoice

- Go to "Accounting / Accounting / Closing / Tax Returns"
- Open the period containing the created invoice
- Mark all lines as "Reviewed"
- Validate the VAT Return

Issue:
In the generated XML, there is a line for grid number "46L",
which should not appear.
Therefore, the XML is rejected by the government.
Same issue with grid number "46T".
These 2 grids are sub-section of grid number "46" and should
not appear in the XML.

Cause:
Previously, they were filtered out, but since this commit
https://github.com/odoo/odoo/commit/17a6117ed88c29b5bc4db0c872bcdbc109a7d98b
the formula has changed from "46L" to "-46L" but the
excluding filter has not been updated.

opw-5344566

closes odoo/enterprise#101353

Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>