[FIX] stock_barcode, sale_stock_renting: allow rental backorders
Steps to reproduce:
- Enable Rental pickings
- Create a rental for a product, for 4 quantity
- Process the delivery
- Open the barcode app and open the reception
- Scan the product once and validate

Issue:
The receipt is validated without issues nor warning, despite being
incomplete.

This is due to a bad mix of two changes:
- #60801, which always sets the rental receipt as return of the delivery
- #48788, which removes the backorder check for returns in barcode

For regular returns made in barcode, it makes sense to avoid the
backorder check, as from here we're processing a full picking return and
we'd have the confirmation pop every time.

However, things are different for rental receipts, as despite them being
set as returns of the delivery, they're proper receipts that need to
handle the partial receipt.

To avoid the issue, rather than removing the backorder check whenever
there's a return linked to the picking, now also checks that there isn't
a rental order linked to the picking.

opw-5265874

closes odoo/enterprise#101690

X-original-commit: 1466a0139ee64ecd059738bc414f6e7e5a9f4354
Related: odoo/odoo#239232
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>