[FIX] l10n_eg_hr_payroll: fix test for 2026
Before this commit, the test `test_get_annual_remaining_leaves_with_allocation`
was checking the balance of an employee with an allocation for 2025.
As no freeze_time was set, the test would fail in 2026 as the allocation
would be invalid and the employee balance would fallback to 0.
Same issue for the test `test_get_annual_remaining_leaves_after_leave_taken`

The PR odoo/enterprise#98223 was targetting the wrong test and thus,
the error was not fixed.

This commit merges all 3 tests in one using freeze_time

runbot error 231559

closes odoo/enterprise#101516

X-original-commit: f9b6800a542fdeeb97ffc74e9630a5f5d2737fd9
Signed-off-by: Abdelrahman Mahmoud (amah) <amah@odoo.com>