[FIX] account_reports: filter from the root reports
For reports that are part of an annual reports, the
_compute_report_option_filter prevent the update of the filters when the report
is added as a section of another report.
In the case of the analytic group by for example, even when the options was
enabled, the filter was not present on the profit and loss (us) since it is a
section of the annual report.

opw-5185296

closes odoo/enterprise#100064

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>