[FIX] ai: add source links to LLM responses as References
This commit allows the LLM now to provide an inline citations within its
claims [SOURCE: chunk_attachment_id1, chunk_attachment_id2] for the
rag source chunks used for each claim.

It also ensures that these citations are transformed into
direct links to the source documents in the final response shown to the
user with a list of sources' names appended to the end of the message.

task-id-5153916

Part-of: odoo/enterprise#101463
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>