[FIX] account_reports: make 'Returns' button of reports branch_allowed
To reproduce the issue:

- Create a company with a branch
- Give the company and its branch different VAT numbers
- Make sure both companies have an opening date, so that they generate returns
- Make both companies active in the company selector
- Open the tax report
- Click on the "Returns" button

===> The following error is raised: "Please select the main company and its branches in the company selector to proceed."

This is because the tax report's options only consider one of the two companies (because they have different VAT numbers). The button is not declared as branch_allowed, so when clicked, it checks whether all the companies of the branch hierachy are in the options => they're not => error.

We can fix this by simply making the "Returns" button branch_allowed.

task-5369592

closes odoo/enterprise#101479

Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>