[IMP] account_reports: extend the refresh period for the return
Previously, the minimum date used for the refresh period in `_try_create_returns_for_fiscal_year` was the fiscal year itself, which caused confusion.
Depending on whether we were at the beginning or the end of the fiscal year, the amount of past data being updated was inconsistent.
The new behavior refreshes a fixed window of one year in the past and one year in the future, making the update range predictable and consistent.

task-5388423

closes odoo/enterprise#101935

X-original-commit: 1ea5f121e1eee847ec67db13dbc4393d194b905a
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>
Signed-off-by: Logan Staelens (lost) <lost@odoo.com>