[FIX] stock_barcode: GS1 failing tour
Issue
=====
During the test tour `test_gs1_package_receipt`, the step 13 adds a new
line for 6 qties of product PRO_GTIN_8. The corresponding line was
successfully created but sometime, the line disappears a short time
after, making following steps unpredictable and making the test fails in
an undeterministic way.

Cause of the issue
==================
The `MainComponent` uses a mutex for the scanned barcode so every
scanned barcodes follow each other, ensuring a deterministic order.
That said, when the record is refreshed (see `_onRefreshState`), we
don't use a mutex, which means it could happen anytime.
Here it is what happening: step 15 we scan a package type, step 17 we
scan a product with quantity, and with a bad timing, it can happen that
the state refresh triggered by the scan of the package type (step 15)
happens just after the new line was created (step 17) and so, the app is
refreshed with a state which doesn't include the new line and is
rendered without it.

Solution
========
To fix this issue, this commit ensures that the `MainComponent` refresh
method also uses the mutex.

This commit also ensures the `BarcodePickingModel` `_processPackageType`
always returns something (since this method is async and can be awaited
by other methods) and modifies the tour `test_gs1_package_receipt` to
use more precise triggers and use helper methods when it's relevant.

Runbot build error: 232683

closes odoo/enterprise#101495

Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>