[FIX] voip: allow regular users to search by extension
When a call is created, VoIP tries to associate it with a contact based,
among other things, one the contact's extension.

However, the extension is stored on the res.users.settings record of
other users, which regular users don't have access to.

This commit updates get_contact_info to perform the search for the
extension in sudo mode, so that regular users also get meaningful
results.

closes odoo/enterprise#101502

Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>