[IMP] account_accountant: Remove expand text feature
This commit removes the expand text feature available
on payment reference field in the bank reconciliation.
Before this commit, a payment reference too long to be
fully displayed in the bank reconciliation view could be
expanded to have the full payment ref.
Now, the user has to unfold the line to get the full
payment ref.

no task id

closes odoo/enterprise#101545

Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>