[FIX] iot: only use webrtc when supported by IoT
Since odoo/odoo#238626, WebRTC support has been removed from the IoT box
in `master` (19.1). However, since IoT boxes can now pair with older
database versions, we need to be able to handle an IoT box that doesn't
support WebRTC.

This commit adds a check to the version field of the IoT box model, and
if it matches the newer format used in 19.1+, WebRTC is skipped.

task-5386539

closes odoo/enterprise#101547

Signed-off-by: Louis Travaux (lotr) <lotr@odoo.com>