[FIX] account_accountant: Optional statement line in chatter
The aim of this commit is putting the bank reco widget
chatter's props "statementLine" as optional.
Since we saved the last state of chatter (open/close),
it might happen that we are opening the chatter without
a statement line selected. This leads to an error in debug
where the props statementLine of our custom chatter is not
filled and leads to a traceback.
By putting the props as optional, we are removing the error.

no task id

closes odoo/enterprise#101548

Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>