[FIX] l10n_be_hr_payroll: Fix holiday pay recovery
Purpose
=======

Normally contracts start and end dates should be configured without
being closed and reopened at each version date.

But, if it is the case, holiday pay recovery could be applied
on older employees because it is considered the employee just joined
the company, and there is an amount to recover.

closes odoo/enterprise#101621

X-original-commit: 3f9a18e0c6513a850e27f7d5a705e1eb6adb1e6e
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>