[FIX] l10_be_intervat: add PyJWT to dependencies
The l10n_be_intervat module needs PyJWT, but it is not listed in external_dependencies.

As we cannot change module dependencies in a stable version, this commit adds a try-except
block regarding the import. In case the library is not installed, an error is logged to inform
the user that they need to install the PyJWT module.

Additionally, this commit updates the import of ImmatureSignatureError to come from jwt.exceptions instead of jwt.

no-task

closes odoo/enterprise#101583

Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>