[FIX] account_reports: add missing account type in balance sheet
In saas-18.3 a new account type was added: "Other Expenses"
These accounts are excluded from the account many2one field to
make it easier to find relevant expense accounts for vendor bills.

Issue: This account type is not included in the Balance Sheet report.

see comment in opw-5269456 related to
opw-5191111

closes odoo/enterprise#101947

X-original-commit: 3c677316c0ef69bc517d713d6776ebd8e5b8ff5d
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>