[FIX] voip: display keypad input cursor in mobile mode
This basically reverts [1] and does what it wanted to do in an hopefully
non-buggy way. The cursor was not appearing anymore because [1] simply
totally removed focus/cursor management in mobile, with the idea of
not showing the mobile keyboard on top of the odoo keyboard.

This commit restores the focus/cursor management and finds a way to hide
the mobile keyboard when it is needed, using the `inputMode` property.

By default, the `<input>` is now rendered with "none" as inputMode,
which should ensure that no update from OWL would open the keyboard. We
also re-force "none" in the function that controls the focus. We only
set to "text" when the user actually engages with the input using his
finger. As soon as the input will change in any other way than using the
mobile keyboard, this will be switched back to "none".

[1]: https://github.com/odoo/enterprise/commit/a06c02e8e163ed052c1c4d25a56bea9b6409abde

task-5366961

closes odoo/enterprise#102728

Signed-off-by: Louis Wicket (wil) <wil@odoo.com>