[FIX] planning: allow single employee user to take multicompany slots
Since #91616, if a user has access to multiple companies but only has
an employee in one, they are unable to assign themselves to a planning
slot from a company other than that of their employee. This was not the
case in previous versions and is causing issues in our internal db.

To restore the previous behavior, any user with access to a company but
only 1 employee will be able to assign themselves to slots of said
company.

opw-5163200

closes odoo/enterprise#101883

X-original-commit: 999c566f20c634c5ddeab2c13d9961173046a6f9
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>